refactor: use named + query in var

+++ src/main/java/com/my/blog/dto/PostUpdateDto.java
@@ -2,7 +2,7 @@ package com.my.blog.dto;
 
 import java.util.List;
 
-public record PostUpdateDto(int id,
+public record PostUpdateDto(Long id,
                             String text,
                             String title,
                             List<String> tags
